configuration of Product model class - tests, controller and services

+++ src/main/java/com/matcss/androidsdungeon/service/ProductService.java
@@ -1,63 +1,68 @@
 package com.matcss.androidsdungeon.service;
 
 import com.matcss.androidsdungeon.interfaces.CRUDServices;
-import com.matcss.androidsdungeon.model.ProductCategory;
-import com.matcss.androidsdungeon.model.Products;
+import com.matcss.androidsdungeon.model.Product;
 import com.matcss.androidsdungeon.repository.ProductRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Slf4j
 @Service
-public class ProductService implements CRUDServices<Products> {
+public class ProductService implements CRUDServices<Product> {
 
-    private final ProductRepository productRepository;
+    @Autowired
+    private ProductRepository productRepository;
+
+    public ProductService() {
+    }
 
     public ProductService(ProductRepository productRepository) {
         this.productRepository = productRepository;
     }
 
-    public List<Products> findAllAvailableProducts(){
+    public List<Product> findAllAvailableProducts(){
         return productRepository
                 .findAll()
                 .stream()
-                .filter(Products::isAvailability)
+                .filter(Product::isAvailability)
                 .collect(Collectors.toList());
     }
 
-    public List<Products> findAllAvailableProductsByCategory(String categoryName){
-        List<Products> products = productRepository
+    public List<Product> findAllAvailableProductsByCategory(String categoryName){
+        List<Product> products = productRepository
                 .findAll()
                 .stream()
+                .filter(Product::isAvailability)
                 .filter(p -> p.getProductCategories()
-                        .stream().anyMatch(pc -> pc.getCategory().getCategory_name().equals(categoryName)))
+                        .stream()
+                        .anyMatch(pc -> pc.getCategory().getCategoryName().equals(categoryName)))
                 .collect(Collectors.toList());
 
         return products;
     }
 
     @Override
-    public List<Products> findAll() {
+    public List<Product> findAll() {
         return productRepository.findAll();
     }
 
     @Override
-    public Products findById(int id) {
+    public Product findById(int id) {
         return productRepository.findByProductId(id);
     }
 
     @Override
-    public Products create(Products obj) {
+    public Product create(Product obj) {
         return productRepository.save(obj);
     }
 
     @Override
-    public Products update(int id, Products obj) {
-        Products product = findById(id);
+    public Product update(int id, Product obj) {
+        Product product = findById(id);
 
         product.setAvailability(obj.isAvailability());
         product.setName(obj.getName());
@@ -69,8 +74,9 @@ public class ProductService implements CRUDServices<Products> {
     }
 
     @Override
-    public void delete(int id) {
-        Products product = findById(id);
-        productRepository.delete(product);
+    public Product delete(int id) {
+        Product product = findById(id);
+        if(product != null) productRepository.delete(product);
+        return product;
     }
 }
